+++ src/gamestate.py
@@ -43,7 +43,7 @@ class GameState:
         self.auto_play: bool = False
         self.motion_model: MotionModels = MotionModels.VECTOR_1
         self.tick_time: int = 0
-        self.cur_ball_x: int = (constants.WIDTH / 2) - (constants.PAD_WIDTH / 2) # used for the auto-play mode that matches paddle pos to the ball pos
+        self.cur_ball_x: float = (constants.WIDTH / 2) - (constants.PAD_WIDTH / 2) # used for the auto-play mode that matches paddle pos to the ball pos
         self.gravity_acc_length: float = constants.WORLD_GRAVITY_ACC
         self.v_gravity_unit: pygame.Vector2 = pygame.Vector2(0.0, 1.0)
         self.v_gravity_acc: pygame.Vector2 = self.v_gravity_unit * self.gravity_acc_length
